PR feedback - moved helper into its own class
