+++ src/main/java/graphql/ExceptionWhileDataFetching.java
@@ -91,11 +91,11 @@ public class ExceptionWhileDataFetching implements GraphQLError {
     @SuppressWarnings("EqualsWhichDoesntCheckParameterClass")
     @Override
     public boolean equals(Object o) {
-        return Helper.equals(this, o);
+        return GraphqlErrorHelper.equals(this, o);
     }
 
     @Override
     public int hashCode() {
-        return Helper.hashCode(this);
+        return GraphqlErrorHelper.hashCode(this);
     }
 }
