+++ src/main/java/graphql/GraphQLError.java
@@ -3,11 +3,8 @@ package graphql;
 
 import graphql.language.SourceLocation;
 
-import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
-
-import static java.util.stream.Collectors.toList;
 
 /**
  * @see <a href="https://facebook.github.io/graphql/#sec-Errors">GraphQL Spec - 7.2.2 Errors</a>
@@ -51,7 +48,7 @@ public interface GraphQLError {
      * @return a map of the error that strictly follows the specification
      */
     default Map<String, Object> toSpecification() {
-        return Helper.toSpecification(this);
+        return GraphqlErrorHelper.toSpecification(this);
     }
 
     /**
@@ -61,64 +58,5 @@ public interface GraphQLError {
         return null;
     }
 
-
-    /**
-     * This little helper allows GraphQlErrors to implement
-     * common things (hashcode/ equals ) and to specification more easily
-     */
-    @SuppressWarnings("SimplifiableIfStatement")
-    class Helper {
-
-        public static Map<String, Object> toSpecification(GraphQLError error) {
-            Map<String, Object> errorMap = new LinkedHashMap<>();
-            errorMap.put("message", error.getMessage());
-            if (error.getLocations() != null) {
-                errorMap.put("locations", locations(error.getLocations()));
-            }
-            if (error.getPath() != null) {
-                errorMap.put("path", error.getPath());
-            }
-            if (error.getExtensions() != null) {
-                errorMap.put("extensions", error.getExtensions());
-            }
-            return errorMap;
-        }
-
-        private static Object locations(List<SourceLocation> locations) {
-            return locations.stream().map(Helper::location).collect(toList());
-        }
-
-        private static Object location(SourceLocation location) {
-            Map<String, Integer> map = new LinkedHashMap<>();
-            map.put("line", location.getLine());
-            map.put("column", location.getColumn());
-            return map;
-        }
-
-        public static int hashCode(GraphQLError dis) {
-            int result = dis.getMessage() != null ? dis.getMessage().hashCode() : 0;
-            result = 31 * result + (dis.getLocations() != null ? dis.getLocations().hashCode() : 0);
-            result = 31 * result + (dis.getPath() != null ? dis.getPath().hashCode() : 0);
-            result = 31 * result + dis.getErrorType().hashCode();
-            return result;
-        }
-
-        public static boolean equals(GraphQLError dis, Object o) {
-            if (dis == o) {
-                return true;
-            }
-            if (o == null || dis.getClass() != o.getClass()) return false;
-
-            GraphQLError dat = (GraphQLError) o;
-
-            if (dis.getMessage() != null ? !dis.getMessage().equals(dat.getMessage()) : dat.getMessage() != null)
-                return false;
-            if (dis.getLocations() != null ? !dis.getLocations().equals(dat.getLocations()) : dat.getLocations() != null)
-                return false;
-            if (dis.getPath() != null ? !dis.getPath().equals(dat.getPath()) : dat.getPath() != null)
-                return false;
-            return dis.getErrorType() == dat.getErrorType();
-        }
-    }
 
 }
