+++ src/main/java/graphql/execution/NonNullableFieldWasNullError.java
@@ -2,6 +2,7 @@ package graphql.execution;
 
 import graphql.ErrorType;
 import graphql.GraphQLError;
+import graphql.GraphqlErrorHelper;
 import graphql.language.SourceLocation;
 
 import java.util.List;
@@ -52,11 +53,11 @@ public class NonNullableFieldWasNullError implements GraphQLError {
     @SuppressWarnings("EqualsWhichDoesntCheckParameterClass")
     @Override
     public boolean equals(Object o) {
-        return Helper.equals(this, o);
+        return GraphqlErrorHelper.equals(this, o);
     }
 
     @Override
     public int hashCode() {
-        return Helper.hashCode(this);
+        return GraphqlErrorHelper.hashCode(this);
     }
 }
