+++ src/main/java/graphql/relay/InvalidCursorException.java
@@ -2,6 +2,7 @@ package graphql.relay;
 
 import graphql.ErrorType;
 import graphql.GraphQLError;
+import graphql.GraphqlErrorHelper;
 import graphql.language.SourceLocation;
 
 import java.util.List;
@@ -31,12 +32,12 @@ public class InvalidCursorException extends RuntimeException implements GraphQLE
     @SuppressWarnings("EqualsWhichDoesntCheckParameterClass")
     @Override
     public boolean equals(Object o) {
-        return Helper.equals(this, o);
+        return GraphqlErrorHelper.equals(this, o);
     }
 
     @Override
     public int hashCode() {
-        return Helper.hashCode(this);
+        return GraphqlErrorHelper.hashCode(this);
     }
 
 }
